requirements.txt
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0
torch>=1.10.0
xgboost>=1.5.0
shap>=0.40.0
networkx>=2.6.0
pyyaml>=6.0
joblib>=1.1.0